requirements.txt
# OpenAI
openai>=1.0.0

# LangChain
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2

# LlamaIndex
llama-index>=0.9.0

# PowerPoint
python-pptx>=0.6.21

# Development Tools
python-dotenv>=1.0.0
pydantic>=2.0.0
typescript>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0